+++ example/lib/cache.dart
@@ -5,7 +5,7 @@ void main() {
   runApp(MyApp());
 }
 
-/// A sample application to demonstrate using SICache
+/// A sample application to demonstrate using ScalableImageCache
 /// (https://github.com/zathras/jovial_svg/issues/6) and to reproduce
 /// race condition issue 7 (https://github.com/zathras/jovial_svg/issues/7).
 
@@ -23,10 +23,15 @@ class MyApp extends StatelessWidget {
 }
 
 class HomePage extends StatelessWidget {
-  final svgs = List.generate(
-      100,
-      (index) => Uri.parse(
-          'https://raw.githubusercontent.com/feathericons/feather/master/icons/activity.svg?x=$index'));
+  static final svgs = List.generate(100,
+      (index) => Uri.parse('https://jovial.com/images/jupiter.svg?x=$index'));
+  // A static cache of the SVG images, so that once they are loaded, they
+  // stay loaded.  Another alternative would be to make HomePage stateful,
+  // and tie the lifetime of the cache to the state.
+  //
+  // For this demo, we make it a little smaller than the length of the list,
+  // so you can see reloading if you scroll all the way back and forth.
+  static final ScalableImageCache svgCache = ScalableImageCache(size: 70);
 
   @override
   Widget build(BuildContext context) {
@@ -41,6 +46,7 @@ class HomePage extends StatelessWidget {
           itemBuilder: (context, index) {
             return GridTile(
               child: ScalableImageWidget.fromSISource(
+                  cache: svgCache,
                   si: ScalableImageSource.fromSvgHttpUrl(svgs[index])),
             );
           }),