+++ CHANGELOG.md
@@ -35,3 +35,6 @@ believed to be representative.
 
 - Address issue 7 (relatively harmless race condition).
 - Add ScalableImageCache (issue 6)
+- Require Flutter 2.5 / SDK 2.14 (Issue 9)
+- Change imageDisposeBugWorkaround default to clean up memory,
+  now that Flutter bug is fixed (Issue 9)